c pasa tests. cropflip asm tambian

+++ codigo/filtros/sepia_c.c
@@ -21,9 +21,9 @@ void sepia_c    (
             bgra_t *p_d = (bgra_t*) &dst_matrix[i][j * 4];
             bgra_t *p_s = (bgra_t*) &src_matrix[i][j * 4];
 			int sumargb = (p_s->r + p_s->g + p_s->b);
-            p_d->r = MIN(0.5*sumargb, 255);
-			p_d->g = MIN(0.3*sumargb, 255);
-			p_d->b = MIN(0.2*sumargb, 255);
+            p_d->r = MIN((int) (0.5*(float) sumargb), 255);
+			p_d->g = MIN((int) (0.3*(float) sumargb), 255);
+			p_d->b = MIN((int) (0.2*(float) sumargb), 255);
             p_d->a = p_s->a;
         }
     }
